+++ docker/osd-dev/manager/entrypoint.sh
@@ -1,12 +1,12 @@
 #!/bin/bash
 
 # Change permissions to certificates
-chmod 500 /etc/filebeat/certs
-chmod 400 /etc/filebeat/certs/*
-chown -R root:root /etc/filebeat/certs
+chmod 500 /etc/server_certs
+chmod 400 /etc/server_certs/*
+chown -R root:root /etc/server_certs
 
-# Start services
-service filebeat start
+# Start service
 /var/ossec/bin/wazuh-control start
 
+# Read logs file
 tail -f /var/ossec/logs/ossec.log