+++ plugins/main/public/utils/known-fields-loader.js
@@ -9,6 +9,17 @@
 // Import index type constants
 import {
   WAZUH_INDEX_TYPE_ALERTS,
+  WAZUH_INDEX_TYPE_ARCHIVES,
+  WAZUH_INDEX_TYPE_EVENTS_ACCESS_MANAGEMENT,
+  WAZUH_INDEX_TYPE_EVENTS_APPLICATIONS,
+  WAZUH_INDEX_TYPE_EVENTS_CLOUD_SERVICES,
+  WAZUH_INDEX_TYPE_EVENTS_CLOUD_SERVICES_AWS,
+  WAZUH_INDEX_TYPE_EVENTS_CLOUD_SERVICES_AZURE,
+  WAZUH_INDEX_TYPE_EVENTS_CLOUD_SERVICES_GCP,
+  WAZUH_INDEX_TYPE_EVENTS_NETWORK_ACTIVITY,
+  WAZUH_INDEX_TYPE_EVENTS_OTHER,
+  WAZUH_INDEX_TYPE_EVENTS_SECURITY,
+  WAZUH_INDEX_TYPE_EVENTS_SYSTEM_ACTIVITY,
   WAZUH_INDEX_TYPE_MONITORING,
   WAZUH_INDEX_TYPE_STATISTICS,
   WAZUH_INDEX_TYPE_STATES_VULNERABILITIES,
@@ -34,6 +45,17 @@ import {
 
 // Import generated known fields JSON files
 import alertsFields from '../../common/known-fields/alerts.json';
+import archivesFields from '../../common/known-fields/archives.json';
+import eventsAccessManagementFields from '../../common/known-fields/events-access-management.json';
+import eventsApplicationsFields from '../../common/known-fields/events-applications.json';
+import eventsCloudServicesFields from '../../common/known-fields/events-cloud-services.json';
+import eventsCloudServicesAWSFields from '../../common/known-fields/events-cloud-services-aws.json';
+import eventsCloudServicesAzureFields from '../../common/known-fields/events-cloud-services-azure.json';
+import eventsCloudServicesGCPFields from '../../common/known-fields/events-cloud-services-gcp.json';
+import eventsNetworkActivityFields from '../../common/known-fields/events-network-activity.json';
+import eventsOtherFields from '../../common/known-fields/events-other.json';
+import eventsSecurityFields from '../../common/known-fields/events-security.json';
+import eventsSystemActivityFields from '../../common/known-fields/events-system-activity.json';
 import vulnerabilitiesFields from '../../common/known-fields/states-vulnerabilities.json';
 import fimFilesFields from '../../common/known-fields/states-fim-files.json';
 import fimRegistriesKeysFields from '../../common/known-fields/states-fim-registries-keys.json';
@@ -68,6 +90,20 @@ export const KnownFieldsByIndexType = {
   [WAZUH_INDEX_TYPE_ALERTS]: alertsFields,
   [WAZUH_INDEX_TYPE_MONITORING]: monitoringFields,
   [WAZUH_INDEX_TYPE_STATISTICS]: statisticsFields,
+  [WAZUH_INDEX_TYPE_ARCHIVES]: archivesFields,
+
+  // Events
+  [WAZUH_INDEX_TYPE_EVENTS_ACCESS_MANAGEMENT]: eventsAccessManagementFields,
+  [WAZUH_INDEX_TYPE_EVENTS_APPLICATIONS]: eventsApplicationsFields,
+  [WAZUH_INDEX_TYPE_EVENTS_CLOUD_SERVICES]: eventsCloudServicesFields,
+  [WAZUH_INDEX_TYPE_EVENTS_CLOUD_SERVICES_AWS]: eventsCloudServicesAWSFields,
+  [WAZUH_INDEX_TYPE_EVENTS_CLOUD_SERVICES_AZURE]:
+    eventsCloudServicesAzureFields,
+  [WAZUH_INDEX_TYPE_EVENTS_CLOUD_SERVICES_GCP]: eventsCloudServicesGCPFields,
+  [WAZUH_INDEX_TYPE_EVENTS_NETWORK_ACTIVITY]: eventsNetworkActivityFields,
+  [WAZUH_INDEX_TYPE_EVENTS_OTHER]: eventsOtherFields,
+  [WAZUH_INDEX_TYPE_EVENTS_SECURITY]: eventsSecurityFields,
+  [WAZUH_INDEX_TYPE_EVENTS_SYSTEM_ACTIVITY]: eventsSystemActivityFields,
 
   // States index types
   [WAZUH_INDEX_TYPE_STATES_VULNERABILITIES]: vulnerabilitiesFields,