+++ plugins/main/server/health-check/index.ts
@@ -1,3 +1,4 @@
 export * from './index-patterns';
 export * from './templates';
 export * from './server-api';
+export * from './types';